+++ src/markdown_renderers.js
@@ -1,5 +1,9 @@
 import log from './log';
 
+const matchAll = require('match-all');
+
+export const APIFY_CODE_TABS = 'apify-code-tabs';
+
 function formatIdTag(idTag) {
     // Get rid of whitespace and random characters
     idTag = idTag.toLowerCase().trim().replace(/[^\w]+/g, '-');
@@ -24,4 +28,58 @@ export const customHeadingRenderer = (text, level, raw) => {
             <h${level} id="${idTag}">${titleText}</h${level}>`;
 
     return headingToReturn;
+};
+
+export const customCodeRenderer = (code) => {
+    if (!code.trim().startsWith('```marked-tabs')) {
+        return false;
+    }
+
+    const matchesIterator = matchAll(code, /<marked-tab header="(?<header>.*?)" lang="(?<lang>.*?)">(?<code>.*?)<\/marked-tab>/sg);
+    const matches = [];
+    let nextMatch = matchesIterator.nextRaw();
+    while (nextMatch) {
+        matches.push(nextMatch);
+        nextMatch = matchesIterator.nextRaw();
+    }
+
+    const inputArg = matches
+        .map((match) => {
+            const { header, lang } = match.groups;
+            let tabCode = match.groups.code;
+
+            let numLeadingSpacesOnFirstLine = 0;
+            while (tabCode.startsWith('\n') || tabCode.startsWith(' ')) {
+                if (tabCode.startsWith('\n')) {
+                    numLeadingSpacesOnFirstLine = 0;
+                } else if (tabCode.startsWith(' ')) {
+                    numLeadingSpacesOnFirstLine += 1;
+                }
+                tabCode = tabCode.substring(1);
+            }
+
+            while (tabCode.endsWith('\n') || tabCode.endsWith(' ')) {
+                tabCode = tabCode.substring(0, tabCode.length - 1);
+            }
+
+            tabCode = tabCode
+                .split('\n')
+                .map((line, index) => {
+                    if (index === 0) {
+                        return line;
+                    }
+
+                    for (let i = 0; i < numLeadingSpacesOnFirstLine; i++) {
+                        if (line.startsWith(' ')) {
+                            line = line.substring(1);
+                        }
+                    }
+                    return line;
+                })
+                .join('\n');
+
+            return `"${header}": { "language": "${lang}", "code: "${tabCode}" }`;
+        })
+        .join(',\n');
+    return `<${APIFY_CODE_TABS} input="{\n${inputArg.replace(/"/g, '\\"')}\n}" />`;
 };
