+++ test/marked.js
@@ -1,10 +1,29 @@
 /* eslint-disable */
 
-import marked from 'marked';
 import { expect } from 'chai';
-import { apifyMarked } from '../src/marked';
+import { overrideMarkedWithCustomSettings } from '../src/marked';
 
-describe('apifyMarked', () => {
+describe('apifyMarked custom renderer works', () => {
+    const marked = overrideMarkedWithCustomSettings();
+
+    it('uses the custom ID if provided', () => {
+        const renderedTitle = marked('# Welcome to Apify {welcome-title-id}');
+        expect(renderedTitle).to.equal(`
+            <h1 id="welcome-title-id">Welcome to Apify</h1>`);
+    });
+
+    it('generates ID from text if no ID provided', () => {
+        const renderedTitle = marked('## Welcome to Apify');
+        expect(renderedTitle).to.eql(`
+            <h2 id="welcome-to-apify">Welcome to Apify</h2>`);
+    });
+
+    it('trims whitespace, inserts dashes between words, converts to lowercase, removes punctuation and trailing dash', () => {
+        const renderedTitle = marked('# Welcome to Apify { #  .Welcome -title-id . - ? -}');
+        expect(renderedTitle).to.eql(`
+            <h1 id="welcome-title-id">Welcome to Apify</h1>`);
+    });
+
     it('correctly parses code block with marked-tabs', () => {
         const markdown = `# Welcome to Apify {welcome-title-id}
             \`\`\`marked-tabs
@@ -24,7 +43,7 @@ describe('apifyMarked', () => {
             \`\`\`
         `;
 
-        expect(apifyMarked(markdown).trim()).to.eql(`<h1 id="welcome-title-id">Welcome to Apify</h1><apify-code-tabs input="{
+        expect(marked(markdown).trim()).to.eql(`<h1 id="welcome-title-id">Welcome to Apify</h1><apify-code-tabs input="{
 \\"NodeJS\\": { \\"language\\": \\"javascript\\", \\"code: \\"console.log('Some JS code');\\" },
 \\"Python\\": { \\"language\\": \\"python\\", \\"code: \\"print('Some python code');
 count = 1
@@ -40,6 +59,7 @@ print('Some python code on next line');\\" },
         ``` \
             console.log(\'Hello World\'); \
         ```"
-        expect(apifyMarked(markdown)).to.eql(marked(markdown));
+        expect(marked(markdown)).to.eql(`
+            <h1 id="test-heading-console-log-hello-world">Test heading         <code>            console.log(&#39;Hello World&#39;);        </code></h1>`);
     });
 });
